requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
outcome==1.1.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.28
trio==0.19.0
trio-websocket==0.9.2
twilio==7.4.0
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
